+++ package.json
@@ -20,7 +20,7 @@
     "migrate": "npx sequelize-cli db:migrate",
     "down": "npx sequelize-cli db:migrate:undo:all",
     "seed": "npx sequelize-cli db:seed:all",
-    "migrate:reset": "npm run down && npm run migrate && npm run seed",
+    "migrate:reset": "cross-env NODE_ENV=test npm run down && cross-env NODE_ENV=test npm run migrate && cross-env NODE_ENV=test npm run seed",
     "test": "cross-env NODE_ENV=test nyc mocha --timeout 40000 ./test/*.test.js --require @babel/register --exit",
     "pretest": "cross-env NODE_ENV=test npm run migrate:reset"
     },
